move models to proper places
