+++ lib/core/services/ecm-user.service.ts
@@ -21,7 +21,7 @@ import { map, catchError } from 'rxjs/operators';
 import { ContentService } from './content.service';
 import { AlfrescoApiService } from './alfresco-api.service';
 import { LogService } from './log.service';
-import { EcmUserModel } from '../userinfo/models/ecm-user.model';
+import { EcmUserModel } from '../models/ecm-user.model';
 import { PersonEntry } from '@alfresco/js-api';
 
 @Injectable({
