+++ lib/core/services/identity-user.service.spec.ts
@@ -42,8 +42,8 @@ import { setupTestBed } from '../testing/setupTestBed';
 import { CoreModule } from '../core.module';
 import { AlfrescoApiService } from './alfresco-api.service';
 import { mockToken } from '../mock/jwt-helper.service.spec';
-import { IdentityUserModel } from '../userinfo/models/identity-user.model';
-import { IdentityRoleModel } from '../userinfo/models/identity-role.model';
+import { IdentityUserModel } from '../models/identity-user.model';
+import { IdentityRoleModel } from '../models/identity-role.model';
 import { AlfrescoApiServiceMock } from '../mock/alfresco-api.service.mock';
 
 describe('IdentityUserService', () => {
