+++ lib/core/services/identity-user.service.ts
@@ -20,9 +20,9 @@ import { Injectable } from '@angular/core';
 import { from, Observable, of, throwError } from 'rxjs';
 import { catchError, map, switchMap } from 'rxjs/operators';
 import { AppConfigService } from '../app-config/app-config.service';
-import { IdentityGroupModel } from '../userinfo/models/identity-group.model';
-import { IdentityRoleModel } from '../userinfo/models/identity-role.model';
-import { IdentityUserModel } from '../userinfo/models/identity-user.model';
+import { IdentityGroupModel } from '../models/identity-group.model';
+import { IdentityRoleModel } from '../models/identity-role.model';
+import { IdentityUserModel } from '../models/identity-user.model';
 import { AlfrescoApiService } from './alfresco-api.service';
 import { JwtHelperService } from './jwt-helper.service';
 import { LogService } from './log.service';
