+++ lib/core/userinfo/public-api.ts
@@ -16,11 +16,11 @@
  */
 
 export * from './components/user-info.component';
-export * from './models/bpm-user.model';
-export * from './models/ecm-user.model';
-export * from './models/identity-group.model';
-export * from './models/identity-user.model';
-export * from './models/identity-role.model';
-export * from './models/identity-group.model';
+export * from '../models/bpm-user.model';
+export * from '../models/ecm-user.model';
+export * from '../models/identity-group.model';
+export * from '../models/identity-user.model';
+export * from '../models/identity-role.model';
+export * from '../models/identity-group.model';
 
 export * from './userinfo.module';
